+++ cmake/nfxGraphicsSources.cmake
@@ -7,6 +7,7 @@ set(NFX_GRAPHICS_SOURCES
     src/gl/mesh/Mesh.cpp
     src/gl/mesh/Primitive.cpp
     src/gl/pipeline/passes/AxesPass.cpp
+    src/gl/pipeline/passes/DirectionalShadowPass.cpp
     src/gl/pipeline/passes/GeometryPass.cpp
     src/gl/pipeline/passes/GridPass.cpp
     src/gl/pipeline/passes/PresentPass.cpp